email_bot/requirements.txt
apiai==1.2.3
certifi==2018.1.18
chardet==3.0.4
Django==2.0
djangorestframework==3.7.7
idna==2.6
numpy==1.14.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
